feat(frontend): Allow for use of proxies for web-pdb sessions
This change makes it so that the frontend looks for the already
in-use hostname and path from the user's session, instead of
assuming that they will always be using the default as set up
by web-pdb. Since I'm using a proxy and docker-compose, I can
make the task of finding a debugger on a per-microservice basis
much easier than if everything is assumed to go to

```
ws://localhost:5555/ws
```

for the websocket connection, and

```
http://localhost:5555/frame-data
```

for the actual pdb session state.

Being able to configure nginx to serve back the per-container
version of the installed web-pdb would be great, that way I
could give each service a unique url that developers could
hit to view that service's debugger running.

Otherwise, they have to run commands like `docker-compose ps`,
figure out what port the web-pdb process got set to on the host
machine, and visit that. Being able to go to something like

```
localhost/debuggers/servicename/
```

would be so much nicer!

As an example, here is one of the current microservice's nginx
entries for setting up a proxy to the api itself, and a couple
extra entries for the web-pdb proxies:

```
location /api/v1/service1/ {
  rewrite ^/api/v1/service1/?(.*) /$1 break;
  proxy_pass http://service1_api:5000;
}

location /debuggers/service1/api/ {
  proxy_pass http://service1_api:5555/;
}

location /frame-data {
  proxy_pass http://service1_api:5555/frame-data;
}

location /ws {
  proxy_pass http://service1_api:5555/ws;
  proxy_http_version 1.1;
  proxy_set_header Upgrade $http_upgrade;
  proxy_set_header Connection "Upgrade";
}
```

With this change, you can instead write:

```
location /api/v1/service1/ {
  rewrite ^/api/v1/service1/?(.*) /$1 break;
  proxy_pass http://service1_api:5000;
}

location /debuggers/service1/api/ {
  rewrite ^/debuggers/service1/api/?(.*) /$1 break;
  proxy_pass http://service1_api:5555;
  proxy_http_version 1.1;
  proxy_set_header Upgrade $http_upgrade;
  proxy_set_header Connection "Upgrade";
}
```