test(docker-compose): Demonstrate what this does with an example
To get this going, install docker, and docker-compose.

Make sure you've already built the frontend part of the app by
following the instructions in the README.md file as this
dockerfile doesn't do any building of the js part of the project.

Run

```
docker-compose up
```

To build and install everything.

Once it comes up, you can visit these three urls to see that the
application can be proxied, and still facilitate the expected
user experience:

```
http://localhost/debuggers/tests/db/
http://localhost/debuggers/tests/db_pm/
http://localhost/debuggers/tests/db_ps/
```